[WIP] Begin adding kwargs.
Unfortuantely got bit by the native function c wrappers of numpy not having introspection, so the code got complex without much to show for it. However, it should be much easier to add more `sklearn.cluster` methods with this refactor. Next, see if there is a unified IO for the kmeans and kmeans_plus_plus for calculating responsibilities.